Delete component InfiniteHits and move it to SearchResults
