+++ src/features/search/atoms/index.tsx
@@ -2,6 +2,7 @@ export * from './FadeScrollingView'
 export * from './FilterSwitch'
 export * from './Hit'
 export * from './NumberOfResults'
+export * from './NoSearchResult'
 export * from './Sections'
 export * from './SelectionLabel'
 export * from './TitleWithCount'
