+++ src/features/search/components/index.tsx
@@ -1,5 +1,4 @@
 export { CalendarPicker } from './CalendarPicker'
-export * from './InfiniteHits'
 export * from './SearchBox'
 export * from './SearchHeader'
 export * from './SearchLandingPage'
